Fix and simplify detection of selected window
Remove the stack of minibuffer selected windows.  Instead consider
(minibuffer-selected-window) the selected window whenever
(selected-window) is a minibuffer.  This seems to be consistent with
what Emacs considers the selected window when choosing the mode line
face.

Also update all mode lines whenever a new window is selected, as
either the previously or the newly selected window would in some cases
not update, resulting in either two or zero windows apparently selected.

Fixes #120