Fix image colors on MacOS
Fixes #54.

Adds a custom option 'powerline-image-apple-rgb', which should be set
to 't' to enable the workaround. When the option is set, powerline's
XPM images are generated using equivalent Apple RGB colors, instead of
the regular sRGB colors.

This is not needed on the Emacs-mac port, or when Apple RGB colors are
used by setting 'ns-use-srgb-colorspace' to nil.