Make Powerline Faces Customisable Without Touching the Mode-Line Face
I introdcued a third font face as powerline-active0 and powerline-inactive0.
Apart from the "center evil" theme (which didn't work for me so I didn't
touch it), all themes have been updated to use this third font face
instead of the mode-line font face (i.e. to configure the left- and
rightmost parts of the mode line). The new faces default to the mode-line
and mode-line-inactive faces respectively, so the default behaviour of
the change is backwards compatible, and existing configurations will
remain unaffected by upgrading to this.

Background: I wanted the left and right parts (where left part holds the
buffer name) to be in a bright colour. To achieve this, I had to set the
mode-line face to have the desired background colour. Undesired side
effect is that "auxiliary" buffers like speedbar, RefTeX toc, etc. become
more visually calling by adopting the bright mode-line background colour.
By not modifying the mode-line face, those auxiliary buffers retain the
default look and feel, whereas buffers in which I am actually editing use
the powerline look and feel and stick out nicely.